migrate uint16 coercion to labview_adapters, rather the the build_return_dict method in the base image analyzer
this allows returning non uint16 image data for further analysis. Note, the naming convention hasn't been changed yet as this is a 'band aid' fix for the moment

+++ ImageAnalysis/image_analysis/base.py
@@ -196,7 +196,8 @@ class ImageAnalyzer:
             """
 
         if return_image is not None:
-            uint_image = return_image.astype(np.uint16)
+            uint_image = return_image   # NOTE, deleted .astype(uint16) but naming is not changed
+                                        # Coerciion to uint16, needed by Labview, is done in labview_adpaters.py
         else:
             uint_image = None
